requirements.txt
asgiref==3.2.10
Django==3.1.1
django-debug-toolbar==2.2
django-filter==2.3.0
psycopg2-binary==2.8.6
python-memcached==1.59
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
